bug: casting removes 0.1-precision

+++ src/gui/SimulatorToolbar.cpp
@@ -31,34 +31,34 @@ SimulatorToolbar::~SimulatorToolbar() {
 
 void SimulatorToolbar::on_sl_velocity_valueChanged( int value ) {
   float val = float( value ) / 10;
-  ui->dsb_velocity->setValue( int(val) );
+  ui->dsb_velocity->setValue( double( val ) );
   emit velocityChanged( val );
 }
 
 void SimulatorToolbar::on_sl_steerangle_valueChanged( int value ) {
   float val = float( value ) / 10;
-  ui->dsb_steerangle->setValue( int(val) );
+  ui->dsb_steerangle->setValue( double( val ) );
   emit steerangleChanged( val );
 }
 
 void SimulatorToolbar::on_sl_frequency_valueChanged( int value ) {
-  ui->dsb_frequency->setValue( value );
+  ui->dsb_frequency->setValue( double( value ) );
   emit frequencyChanged( value );
 }
 
 void SimulatorToolbar::on_dsb_velocity_valueChanged( double val ) {
-  ui->sl_velocity->setValue( int(val * 10) );
-  emit velocityChanged( float(val) );
+  ui->sl_velocity->setValue( int( val * 10 ) );
+  emit velocityChanged( float( val ) );
 }
 
 void SimulatorToolbar::on_dsb_steerangle_valueChanged( double val ) {
-  ui->sl_steerangle->setValue( int(val * 10) );
-  emit steerangleChanged(float( val) );
+  ui->sl_steerangle->setValue( int( val * 10 ) );
+  emit steerangleChanged( float( val ) );
 }
 
 void SimulatorToolbar::on_dsb_frequency_valueChanged( double val ) {
-  ui->sl_frequency->setValue( int(val) );
-  emit frequencyChanged(int( val) );
+  ui->sl_frequency->setValue( int( val ) );
+  emit frequencyChanged( int( val ) );
 }
 
 void SimulatorToolbar::on_pb_frequencyReset_clicked() {
